added methods to get nodes/flows
